[feature] bottom navigation 구현, home screen 파일 생성
