+++ gradle/libs.versions.toml
@@ -10,6 +10,8 @@ hiltCompiler = "1.2.0"
 
 # data layer
 material3AdaptiveNavigationSuite = "1.0.0-alpha07"
+materialVersion = "1.6.8"
+navigationCompose = "2.7.7"
 retrofit2 = "2.9.0"
 okhttp = "4.10.0"
 ktx = "1.13.1"
@@ -36,7 +38,11 @@ compose-glide = "1.4.7"
 [libraries]
 # {groupId}:{artifactId}:{version}
 
+androidx-material = { module = "androidx.compose.material:material", version.ref = "materialVersion" }
 androidx-material3-adaptive-navigation-suite = { module = "androidx.compose.material3:material3-adaptive-navigation-suite", version.ref = "material3AdaptiveNavigationSuite" }
+androidx-navigation-compose = { module = "androidx.navigation:navigation-compose", version.ref = "navigationCompose" }
+androidx-navigation-fragment-ktx = { module = "androidx.navigation:navigation-fragment-ktx", version.ref = "navigationCompose" }
+androidx-navigation-ui-ktx = { module = "androidx.navigation:navigation-ui-ktx", version.ref = "navigationCompose" }
 material3 = { module = "androidx.compose.material3:material3", version.ref = "composeMaterial" }
 dagger-hilt-compiler = { module = "com.google.dagger:hilt-compiler", version.ref = "hilt" }
 hilt = { group = "com.google.dagger", name = "hilt-android", version.ref = "hilt" }
@@ -86,3 +92,4 @@ android-application = { id = "com.android.application", version.ref = "androidGr
 android-library = { id = "com.android.library", version.ref = "androidGradlePlugin"}
 kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin"}
 hilt-android = { id = "com.google.dagger.hilt.android", version.ref = "hilt" }
+safeargs-kotlin = { id = 'androidx.navigation.safeargs.kotlin', version.ref = "navigationCompose"}
