+++ presentation/build.gradle
@@ -2,6 +2,7 @@ plugins {
     alias(libs.plugins.android.library)
     alias(libs.plugins.kotlin.android)
     alias(libs.plugins.hilt.android)
+    alias(libs.plugins.safeargs.kotlin)
 }
 
 apply plugin: 'com.google.dagger.hilt.android'
@@ -81,6 +82,11 @@ dependencies {
     implementation(libs.coil)
     implementation(libs.glide)
     implementation(libs.compose.glide)
+
+    implementation libs.androidx.navigation.fragment.ktx
+    implementation libs.androidx.navigation.ui.ktx
+    implementation libs.androidx.navigation.compose
+    implementation libs.androidx.material
 
     implementation project(path: ':domain')
 }
